Fixes a load time infinite loop with AOT compiled code
The issue is in the redefining of the translate function. When running
without AOT, that kind of var changing is ok, but AOT doesn't allow
it. This commit switches away from that and replaces it with two
app-name functions that will have the translated Metabase string OR
the user defined application-name