Remove row level permissions overly restrictive perms checking
The row level permissions middleware previously only allowed passing
through the middleware if full data access or segmented access was
found. With the new spaces and collections code, users can have view
access to a collection which is neither segmented or full data access.

This commit removes that overly restrictive permissions checking that
is really just redundant since this permissions checking is already
applied in the routing layer. With this commit we're checking for
permissions just to determine whether or not we need to apply
segmented permissions.

Fixes #27