Fix segmented permissions for Card PUTs
If the user changed the query associated with the card and had valid
segmented permissions, the user would falsely get a 403.

This adds the same check on PUT that is used for POST. It also hids
the segmented and full table permissions check behind a function so as
to not sprinkle segmented perms through-out the API layer.