Fixes row level perms breaking queries on nested cards
Specifically there was a check in the predicate deciding to apply row
level permissions that only looked for source tables and didn't handle
the source table being a card of the form `card__<id>`. This fixes
that bug and adds a test to the row level permissions test suite to
ensure that works.