Add a column level permissions check and GTAP FK test
The column level permissions check is in place to prevent a potential
security issue where a user attempts to breakout by a column they
don't have access to. With the nested query FK related code, we would
normally include that in the projection. We don't want that in this
case as the user doesn't have access to that.

This also creates a specific GTAP with FK test. The test is very
similar to the nested MBQL with FK tests but this ensures that the
same functionality is available for GTAPs that automatically swap the
original table with a row level restricted query.