Ensure connections are closed when a db is deleted [ci drivers]
This can cause issues, especially in test code that creates and
destroys many databases throughout the test suite. The connections are
left open and eventually the database runs out of connections, causing
the test run to fail.

This commit also makes a few other minor test related changes such as
adding a `table-name` similar to our `id` function that returns the
correct table name for the given database.