Don't accept valid Google Auth tokens not meant for us
With checking that the `aud`ience field in the validated token info, it
is possible to reuse the Google Auth token for another site/app to log
into Metabase. This could allow authentication bypass for Metabase user
accounts, under certain circumstances.

The `google-auth-token-info` function was refactored to make it more
testable. This also required that the HTTP request to retrieve the token
info be moved to where `google-auth-token-info` is called.